siihdmi: fix wake from unblank after powerdown
If the device is put into D2 state, we need to restore the device to D0 and then
re-setup the display output.  This is only a problem when we receive a
FB_BLANK_POWERDOWN followed by a FB_BLANK_UNBLANK.  Because this can cause
delays during boot, asychronously rescan the device via the hot-plug codepath
upon an unblank.

Signed-off-by: Saleem Abdulrasool <compnerd@compnerd.org>