HACK: fix -marm and -march options to counteract compiler defaults
This integrates an upstream patch (91802a8e) entitled;

"ARM: pass -marm to gcc by default for both C and assembler"

Which essentially fixes the case with a toolchain that is built to
default to -mthumb (as is the case for Ubuntu/Linaro toolchains)
and solves issues with the assembler getting confused when presented
with -mno-thumb-interwork (since it then never generates IT blocks)
and fixes the build.

We also fix an annoying warning in a hacky way; upstream suggests fixing
this in a patch entitled;

"ARM: compressed/head.S: work around new binutils warning"

But instead of littering head.S with new statements to get around
things, since the Slimbook has a Cortex-A8 or better we may as well
just build it for -march=armv7-a instead of -march=all.

Signed-off-by: Matt Sealey <matt@genesi-usa.com>